+++ apps/admin-web/src/features/marketplace/hooks/useCreateScraperConfig.ts
@@ -1,11 +1,11 @@
 import { useRef } from 'react';
+import { showRetryToast } from '@shared/ui/mutation-toast';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
 import type { CreateScraperConfigPayload } from '../services/revolico-service';
 import { revolicoService } from '../services/revolico-service';
 
-import { showRetryToast } from './mutation-toast';
 import { marketplaceKeys } from './query-keys';
 
 /**
