+++ apps/admin-web/src/features/marketplace/hooks/useDeleteSchedule.ts
@@ -1,10 +1,10 @@
 import { useRef } from 'react';
+import { showRetryToast } from '@shared/ui/mutation-toast';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
 import { schedulesService } from '../services/schedules-service';
 
-import { showRetryToast } from './mutation-toast';
 import { marketplaceKeys } from './query-keys';
 
 /**
