+++ apps/admin-web/src/features/marketplace/hooks/useTriggerScrapingJob.ts
@@ -1,11 +1,10 @@
 import { useRef } from 'react';
+import { showRetryToast } from '@shared/ui/mutation-toast';
 import { useMutation } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
 import type { ScrapeJobPayload } from '../services/revolico-service';
 import { revolicoService } from '../services/revolico-service';
-
-import { showRetryToast } from './mutation-toast';
 
 /**
  * Triggers a manual Revolico scraping job.
