+++ apps/admin-web/src/features/rules/hooks/useCreateRule.ts
@@ -1,7 +1,7 @@
+import { showRetryToast } from '@shared/ui/mutation-toast';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 
-import { showRetryToast } from '../../marketplace/hooks/mutation-toast';
 import type { CreateRulePayload } from '../services/rules-service';
 import { rulesService } from '../services/rules-service';
